refactor and instructions
